# specs/001-face-checkin-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FaceCheck API
  version: 1.0.0
  description: |
    Spring Boot 单体后端对 Flutter App 暴露的首版接口契约。
    PostgreSQL 是唯一业务事实源；华为云 FRS 仅作为托管识别能力；
    华为云 OBS 仅作为图片对象存储。
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/admin/sessions:
    get:
      tags: [AdminSessions]
      summary: List attendance sessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummaryList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [AdminSessions]
      summary: Create an attendance session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/sessions/{sessionId}:
    put:
      tags: [AdminSessions]
      summary: Update an attendance session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAttendanceSessionRequest'
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/sessions/{sessionId}/publish:
    post:
      tags: [AdminSessions]
      summary: Publish a session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/sessions/{sessionId}/close:
    post:
      tags: [AdminSessions]
      summary: Close a session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/sessions/{sessionId}/cancel:
    post:
      tags: [AdminSessions]
      summary: Cancel a session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAttendanceSessionSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/sessions/{sessionId}/qr-token:
    get:
      tags: [AdminSessions]
      summary: Read current session QR token
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Current QR token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQrTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/sessions/{sessionId}/qr-token/reset:
    post:
      tags: [AdminSessions]
      summary: Regenerate current session QR token
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: QR token reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseQrTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiResponseAttendanceSessionSummary:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/AttendanceSessionSummary'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ApiResponseAttendanceSessionSummaryList:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceSessionSummary'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ApiResponseQrTokenResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/QrTokenResponse'
        error:
          nullable: true
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    CreateAttendanceSessionRequest:
      type: object
      required: [name, startTime, endTime]
      properties:
        name:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        lateAfterTime:
          type: string
          format: date-time
          nullable: true
    UpdateAttendanceSessionRequest:
      allOf:
        - $ref: '#/components/schemas/CreateAttendanceSessionRequest'
    AttendanceSessionSummary:
      type: object
      required: [sessionId, name, startTime, endTime, status, createdAt]
      properties:
        sessionId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        lateAfterTime:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CLOSED, CANCELED]
        qrTokenVersion:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    QrTokenResponse:
      type: object
      required: [sessionId, qrToken, qrContent]
      properties:
        sessionId:
          type: string
          format: uuid
        qrToken:
          type: string
        qrContent:
          type: string
          description: App deep link or encoded payload content
...
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorDetail'
        traceId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string